+++ IfcGeometryResource/IfcCartesianTransformationOperator2DnonUniform.java
@@ -5,7 +5,7 @@ package buildingsmart.ifc
 {
 	public class IfcCartesianTransformationOperator2DnonUniform extends IfcCartesianTransformationOperator2D
 	{
-		private double Scale2;
-		private double Scl2;
+		private REAL Scale2;
+		private REAL Scl2;
 	}
 }
